# specs/openapi.yaml
openapi: 3.0.3
info:
  title: HouseFlow API
  description: |
    API pour l'application HouseFlow - Suivi d'entretien de maison.

    ## MVP Features
    - Authentification (inscription, connexion, déconnexion)
    - Gestion des maisons (CRUD)
    - Gestion des appareils (CRUD)
    - Gestion des types d'entretien (CRUD)
    - Logging des entretiens effectués
    - Calcul des scores de progression

    ## Conventions
    - **Erreurs** : ProblemDetails (`application/problem+json`, schéma `ErrorResponse`) avec un
      `code` machine quand l'interface doit réagir (voir `ErrorResponse.code`).
    - **403 / 404** : 404 (`not_found`) si la ressource — ou la maison qui la contient — n'existe
      pas ; 403 (`forbidden`) si elle existe mais que l'appelant n'a pas le rôle requis (R5).
    - **Dates calendaires** (date d'entretien, échéance, date d'installation) : un jour du
      calendrier Europe/Paris, transporté comme **minuit UTC de ce jour**
      (`yyyy-MM-ddT00:00:00Z`). Les réponses l'envoient toujours sous cette forme
      (`format: date-time`) ; les requêtes acceptent `yyyy-MM-dd` ou `yyyy-MM-ddT00:00:00Z`
      (`format: date`) — seule la partie date est prise en compte, jamais un instant local.
  version: 1.0.0
  contact:
    name: HouseFlow Team
servers:
  - url: http://localhost:5203/api/v1
    description: Local Development Server
  - url: https://api.houseflow.app/v1
    description: Production Server
...
paths:
...
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Rafraîchir le token
      description: >-
        Échange le refresh token (cookie HttpOnly `refreshToken`) contre un nouveau
        access token. Le refresh token est tourné à chaque appel ; présenter un token
        déjà tourné (hors fenêtre de grâce de 30 s) révoque toute la famille de tokens
        issue du même login (détection de vol).
      tags: [Auth]
      responses:
        '200':
          description: Token rafraîchi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: |
            Refresh token absent, invalide, expiré ou réutilisé (`invalid_refresh_token`), ou
            compte sous limitation du traitement, RGPD Art. 18 (`account_restricted`).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/me:
    get:
      operationId: getMyProfile
      summary: Profil de l'utilisateur connecté
      description: |
        Retourne le profil de l'utilisateur authentifié, y compris l'état de son
        consentement (RGPD Art. 15 — droit d'accès, données de profil).
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profil utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Non authentifié
    put:
      operationId: updateMyProfile
      summary: Modifier son profil (droit de rectification)
      description: |
        RGPD Art. 16 — permet à l'utilisateur de rectifier ses données d'identité
        (prénom, nom, email). L'email doit rester unique.
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profil mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Données invalides
        '401':
          description: Non authentifié
        '409':
          description: Email déjà utilisé par un autre compte (`email_taken`)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteMyAccount
      summary: Supprimer son compte (droit à l'effacement)
      description: |
        RGPD Art. 17 — suppression définitive et immédiate du compte et de toutes les
        données personnelles associées :
        - maisons dont l'utilisateur est propriétaire : transférées au collaborateur
          non-locataire le plus ancien (CollaboratorRW, sinon CollaboratorRO) s'il en
          existe un, sinon supprimées avec tout leur contenu (appareils, entretiens,
          membres, invitations en cascade) ;
        - adhésions aux maisons d'autres utilisateurs (retrait du membre uniquement) ;
        - refresh tokens et clés API (révocation immédiate de toutes les sessions) ;
        - invitations créées par l'utilisateur ;
        - journaux d'audit : anonymisés (userId → null, username → "deleted-user",
          identifiant du compte → "deleted", IP/user agent/valeurs → supprimés) et
          conservés pour la traçabilité. Une fois le lien avec la personne rompu,
          l'enregistrement n'est plus une donnée personnelle (considérant 26 ; avis
          WP216) ; la finalité de sécurité repose sur l'Art. 6(1)(f).
        Confirmation par ressaisie du mot de passe obligatoire.
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAccountRequest'
      responses:
        '204':
          description: Compte supprimé. Le cookie refreshToken est effacé.
        '400':
          description: Mot de passe incorrect ou requête invalide
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Non authentifié
  /users/me/export:
    get:
      operationId: exportMyData
      summary: Exporter toutes ses données (droit d'accès et portabilité)
      description: |
        RGPD Art. 15 (accès) et Art. 20 (portabilité) — retourne une copie complète
        des données personnelles de l'utilisateur dans un format structuré, couramment
        utilisé et lisible par machine.
        - `format=json` (défaut) : un document JSON (`UserDataExport`),
          `Content-Disposition: attachment; filename="houseflow-data-export-{yyyy-MM-dd}.json"`.
        - `format=csv` : une archive ZIP contenant un fichier CSV (UTF-8, séparateur `,`)
          par catégorie (profile, houses, devices, maintenance_types, maintenance_instances,
          memberships, invitations, api_keys, sessions, audit_logs),
          `Content-Disposition: attachment; filename="houseflow-data-export-{yyyy-MM-dd}.zip"`.
        Les secrets techniques (hash du mot de passe, valeurs des tokens, hash des clés API)
        ne sont jamais exportés. Limité à 1 export par heure et par utilisateur (429).
        Chaque export est journalisé (audit "DataExport").
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: format
          in: query
          required: false
          description: "`json` (défaut) ou `csv`"
          schema:
            type: string
            default: json
      responses:
        '200':
          description: Export des données
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDataExport'
            application/zip:
              schema:
                type: string
                format: binary
        '400':
          description: Format non pris en charge (ni `json` ni `csv`) — le quota horaire n'est pas consommé
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Non authentifié
        '429':
          description: Un export a déjà été effectué il y a moins d'une heure (`export_rate_limited`, header Retry-After)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via /auth/login ou /auth/register
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clé API générée depuis les paramètres utilisateur
...
  schemas:
...
    ErrorResponse:
      type: object
      description: |
        Erreur au format ProblemDetails (RFC 9457, `application/problem+json`). Le frontend
        branche sur `status` et `code`, jamais sur le texte (`detail` est en anglais, non
        destiné à l'affichage). Les 400 de validation de modèle ajoutent `errors`.
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
        detail:
          type: string
          nullable: true
          description: Message technique (anglais)
        code:
          type: string
          nullable: true
          description: |
            Code machine stable, quand l'interface doit réagir spécifiquement :
            `invalid_credentials`, `account_restricted`, `invalid_refresh_token`, `email_taken`,
            `export_rate_limited`, `invitation_invalid`, `invitation_email_mismatch`,
            `invitation_already_pending`, `invitation_limit_reached`, `already_member`,
            `own_invitation`, `forbidden`, `not_found`.
        errors:
          type: object
          nullable: true
          additionalProperties:
            type: array
            items:
              type: string
          description: Erreurs de validation par champ (400 de validation uniquement)

    # ==========================================================================
    # AUTH
    # ==========================================================================
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: Token pour rafraîchir l'access token
        expiresIn:
          type: integer
          description: Durée de validité en secondes
        user:
          $ref: '#/components/schemas/User'
        joinedHouseId:
          type: string
          format: uuid
          nullable: true
          description: Inscription avec `invitationToken` uniquement — maison rejointe (P03 → P09)
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        theme:
          type: string
          enum: [light, dark, system]
          default: system
          description: Préférence de thème utilisateur
        language:
          type: string
          enum: [fr, en]
          default: fr
          description: Langue préférée de l'utilisateur
        consentRequired:
          type: boolean
          default: false
          description: |
            RGPD — `true` si l'utilisateur doit (ré)accepter la politique de
            confidentialité / CGU en vigueur (compte créé avant l'introduction du
            consentement, ou nouvelle version de la politique). Le frontend affiche
            alors une bannière de re-consentement (POST /users/me/consent).
        isAdmin:
          type: boolean
          default: false
          description: Administrateur de la plateforme (accès à l'interface d'administration)

    # ==========================================================================
    # HOUSES
    # ==========================================================================
...
    UserProfile:
      type: object
      required: [id, firstName, lastName, email, theme, language, createdAt, consentRequired]
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        theme:
          type: string
          enum: [light, dark, system]
        language:
          type: string
          enum: [fr, en]
        createdAt:
          type: string
          format: date-time
        consentGivenAt:
          type: string
          format: date-time
          nullable: true
          description: Date d'acceptation de la politique de confidentialité / CGU (null pour les comptes antérieurs)
        consentPolicyVersion:
          type: string
          nullable: true
          description: Version de la politique acceptée (format date ISO, ex. "2026-09-11")
        consentRequired:
          type: boolean
          description: true si l'utilisateur doit (ré)accepter la politique en vigueur
    UpdateProfileRequest:
      type: object
      required: [firstName, lastName, email]
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 255
    DeleteAccountRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 1
          description: Mot de passe actuel, ressaisi pour confirmer la suppression
...
    UserDataExport:
      type: object
      description: |
        Copie complète des données personnelles de l'utilisateur (Art. 15 + 20).
        Structure stable, destinée à être lisible par machine et importable ailleurs.
      properties:
        exportedAt:
          type: string
          format: date-time
        formatVersion:
          type: string
          example: "1.0"
        profile:
          type: object
          description: id, email, firstName, lastName, createdAt, updatedAt
        preferences:
          type: object
          description: theme, language
        consent:
          type: object
          description: consentGivenAt, consentPolicyVersion
        houses:
          type: array
          description: Maisons possédées, avec devices → maintenanceTypes → maintenanceInstances imbriqués
          items:
            type: object
        memberships:
          type: array
          description: Maisons partagées avec l'utilisateur (houseName, role, canLogMaintenance, canViewCosts, since)
          items:
            type: object
        invitationsSent:
          type: array
          items:
            type: object
        invitationsReceived:
          type: array
          items:
            type: object
        apiKeys:
          type: array
          description: name, prefix, scope, createdAt, lastUsedAt, revokedAt (jamais le hash)
          items:
            type: object
        sessions:
          type: array
          description: Refresh tokens — createdAt, expiresAt, createdByIp, revokedAt, revokedByIp, reasonRevoked (jamais la valeur du token)
          items:
            type: object
        auditLogs:
          type: array
          description: Actions effectuées par l'utilisateur — timestamp, action, entityType, entityId, changedProperties, ipAddress, userAgent
          items:
            type: object
        information:
          type: object
          description: |
            Volet « métadonnées » de l'export (Art. 15(1)(a) à (h) et 15(2)), sans lequel
            un export n'est qu'un extrait de base de données : controller, contactEmail,
            policyVersion, purposes, dataCategories, legalBases, recipients, retention,
            rights, supervisoryAuthorities, dataSource, internationalTransfers,
            automatedDecisionMaking et portabilityScope (sections relevant de l'Art. 20).
            Les textes sont fournis en anglais et en français (`{ en, fr }`).

    # ==========================================================================
    # ADMIN
    # ==========================================================================
...